LayoutTests/imported/w3c/web-platform-tests/dom/observable/tentative/observable-take.any-expected.txt


PASS take(): Takes the first N values from the source observable, then completes
PASS take(): Forwards complete()s that happen before the take count is met, and unsubscribes from source Observable
PASS take(): Should forward errors from the source observable
PASS take(): take(0) should not subscribe to the source observable, and should return an observable that immediately completes
PASS take(): Negative count is treated as maximum value
